+++ llama/llama_cpp.py
@@ -48,7 +48,7 @@ os.environ['TRANSFORMERS_NO_ADVISORY_WARNINGS'] = os.getenv('TRANSFORMERS_NO_ADV
 
 import atexit
 from enum import Enum
-from threading import Lock
+# from threading import Lock
 from typing import TypeAlias
 from weakref import WeakKeyDictionary
 
@@ -161,7 +161,26 @@ class ggml_type(Enum):
     # IQ4_NL_8_8 = 38
     # COUNT   = 39
 
-lock = Lock()
+# lock = Lock()
+
+class DummyLock:
+    def __init__(self):
+        pass
+
+    def acquire(self, blocking=True, timeout=-1):
+        return True
+
+    def release(self):
+        pass
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        pass
+
+
+lock = DummyLock()
 
 
 # Set callback for all future logging events.